requirements.txt
Django==3.1.7
django-extensions==3.1.1
Pillow==8.1.2
